[Fix] Frame processing/Episode reset problems
Resolved a problem where the first fame wouldn't get processed.
Resolved a problem where the environment constantly resets.
